add security to the change request "get by id" methods
OCD-2159
